Rename baseline to fldsme to match Flower baseline naming convention
Fixes bot review: baseline tooling expects folder name to match
Python package name. Renamed from fl-dsme-iot/fl_dsme_iot to
fldsme/fldsme following the convention of fedprox, fedavgm, fednova etc.
